+++ contracts/fund/src/events.rs
@@ -7,7 +7,7 @@ use cosmwasm_std::{Coin, Event};
 use cw2::ContractVersion;
 
 pub fn event_withdraw(user: &str, amount_withdrawn: Coin) -> Event {
-    Event::new("withdraw_and_swap").add_attributes([
+    Event::new("withdraw").add_attributes([
         ("version", CONTRACT_VERSION),
         ("contract", CONTRACT_NAME),
         ("user", user),
@@ -15,12 +15,22 @@ pub fn event_withdraw(user: &str, amount_withdrawn: Coin) -> Event {
     ])
 }
 
-pub fn event_repay(user: String, token_in: Coin) -> Event {
+pub fn event_repay(user: &str, token_in: Coin) -> Event {
     Event::new("repay").add_attributes([
         ("version", CONTRACT_VERSION),
         ("contract", CONTRACT_NAME),
-        ("user", &user),
+        ("user", user),
         ("amount_repayed", &token_in.to_string()),
+    ])
+}
+
+pub fn event_sudo(to: &str, from: &str, amount: &Coin) -> Event {
+    Event::new("sudo").add_attributes([
+        ("version", CONTRACT_VERSION),
+        ("contract", CONTRACT_NAME),
+        ("to", to),
+        ("from", from),
+        ("amount", &amount.to_string()),
     ])
 }
 
@@ -104,6 +114,11 @@ pub fn event_redeem(
             &fee1.map_or("-".to_string(), |coin| coin.to_string()),
         ),
     ])
+}
+
+pub fn event_register_sudo() -> Event {
+    Event::new("register_sudo")
+        .add_attributes([("version", CONTRACT_VERSION), ("contract", CONTRACT_NAME)])
 }
 
 pub fn event_migrate(version: &str, name: &str, contract_version: ContractVersion) -> Event {