+++ contracts/fund/src/lib.rs
@@ -6,10 +6,12 @@ pub mod handle;
 pub mod helpers;
 pub mod math;
 pub mod messages;
+pub mod process;
 pub mod queries;
 pub mod query;
 pub mod reply;
 pub mod state;
+pub mod sudo;
 
 #[cfg(test)]
 mod testing;