+++ contracts/fund/src/testing/integration/redeem.rs
@@ -1,5 +1,6 @@
 use cosmwasm_std::{assert_approx_eq, coin, coins, Uint128};
 use interface::fund::StateResponse;
+use neutron_std::types::osmosis::tokenfactory::WhitelistedHook;
 use neutron_test_tube::{Account, Module, Wasm};
 use testing::setup::{TestEnv, BASE_DENOM, QUOTE_DENOM};
 
@@ -253,19 +254,21 @@ fn test_redeem_end_to_end_with_repayment() {
             &env.controller,
         )
         .unwrap();
+        let latest_block_time = env.app.get_block_timestamp();
+
+        let expected_state = StateResponse {
+            is_open: true,
+            is_paused: false,
+            total_staked_tokens: Uint128::new(300_000_000),
+            total_withdrawn_tokens: coins(150_000_000, BASE_DENOM),
+            last_pause: block_time,
+            last_claim: latest_block_time,
+            pending_management_fees: vec![],
+        };
+
+        let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
+        assert_eq!(state, expected_state);
     }
-
-    let expected_state = StateResponse {
-        is_open: true,
-        is_paused: false,
-        total_staked_tokens: Uint128::new(300_000_000),
-        total_withdrawn_tokens: coins(150_000_000, BASE_DENOM),
-        last_pause: block_time,
-        last_claim: block_time,
-    };
-
-    let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
-    assert_eq!(state, expected_state);
 
     // Controller repays - with profit
     {
@@ -287,6 +290,7 @@ fn test_redeem_end_to_end_with_repayment() {
         );
         env.redeem_fund(&wasm, &vault_addr, redeem.clone(), &env.traders[0])
             .unwrap();
+        let latest_block_time = env.app.get_block_timestamp();
 
         let expected_state = StateResponse {
             is_open: true,
@@ -294,7 +298,8 @@ fn test_redeem_end_to_end_with_repayment() {
             total_staked_tokens: Uint128::new(200_000_000),
             total_withdrawn_tokens: coins(0u128, BASE_DENOM),
             last_pause: block_time,
-            last_claim: block_time,
+            last_claim: latest_block_time,
+            pending_management_fees: vec![],
         };
 
         let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
@@ -355,19 +360,21 @@ fn test_redeem_end_to_end_with_partial_repayment() {
             &env.controller,
         )
         .unwrap();
+
+        let latest_block_time = env.app.get_block_timestamp();
+        let expected_state = StateResponse {
+            is_open: true,
+            is_paused: false,
+            total_staked_tokens: Uint128::new(300_000_000),
+            total_withdrawn_tokens: coins(150_000_000, BASE_DENOM),
+            last_pause: block_time,
+            last_claim: latest_block_time,
+            pending_management_fees: vec![],
+        };
+
+        let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
+        assert_eq!(state, expected_state);
     }
-
-    let expected_state = StateResponse {
-        is_open: true,
-        is_paused: false,
-        total_staked_tokens: Uint128::new(300_000_000),
-        total_withdrawn_tokens: coins(150_000_000, BASE_DENOM),
-        last_pause: block_time,
-        last_claim: block_time,
-    };
-
-    let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
-    assert_eq!(state, expected_state);
 
     // Controller repays - with no profit
     {
@@ -385,6 +392,7 @@ fn test_redeem_end_to_end_with_partial_repayment() {
         );
         env.redeem_fund(&wasm, &vault_addr, redeem.clone(), &env.traders[0])
             .unwrap();
+        let latest_block_time = env.app.get_block_timestamp();
 
         let expected_state = StateResponse {
             is_open: true,
@@ -392,7 +400,8 @@ fn test_redeem_end_to_end_with_partial_repayment() {
             total_staked_tokens: Uint128::new(200_000_000),
             total_withdrawn_tokens: coins(50_000_000, BASE_DENOM), // Shortfall in repayment
             last_pause: block_time,
-            last_claim: block_time,
+            last_claim: latest_block_time,
+            pending_management_fees: vec![],
         };
 
         let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
@@ -455,28 +464,31 @@ fn test_redeem_end_to_end_with_repayment_multiple_denom() {
     }
 
     // Controller withdraws
+    {
+        let withdraw_amount_base = coin(150_000_000, BASE_DENOM);
+        let withdraw_amount_quote = coin(50_000_000, QUOTE_DENOM);
+        env.withdraw_strategy(
+            &wasm,
+            &strategy_addr,
+            vec![withdraw_amount_base.clone(), withdraw_amount_quote.clone()],
+            &env.controller,
+        )
+        .unwrap();
+        let latest_block_time = env.app.get_block_timestamp();
 
-    let withdraw_amount_base = coin(150_000_000, BASE_DENOM);
-    let withdraw_amount_quote = coin(50_000_000, QUOTE_DENOM);
-    env.withdraw_strategy(
-        &wasm,
-        &strategy_addr,
-        vec![withdraw_amount_base.clone(), withdraw_amount_quote.clone()],
-        &env.controller,
-    )
-    .unwrap();
+        let expected_state = StateResponse {
+            is_open: true,
+            is_paused: false,
+            total_staked_tokens: Uint128::new(425_000_000),
+            total_withdrawn_tokens: vec![withdraw_amount_quote, withdraw_amount_base],
+            last_pause: block_time,
+            last_claim: latest_block_time,
+            pending_management_fees: vec![],
+        };
 
-    let expected_state = StateResponse {
-        is_open: true,
-        is_paused: false,
-        total_staked_tokens: Uint128::new(362_500_000),
-        total_withdrawn_tokens: vec![withdraw_amount_base, withdraw_amount_quote],
-        last_pause: block_time,
-        last_claim: block_time,
-    };
-
-    let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
-    assert_eq!(state, expected_state);
+        let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
+        assert_eq!(state, expected_state);
+    }
 
     // Controller repays - with profit
     {
@@ -505,14 +517,16 @@ fn test_redeem_end_to_end_with_repayment_multiple_denom() {
         );
         env.redeem_fund(&wasm, &vault_addr, redeem.clone(), &env.traders[0])
             .unwrap();
+        let latest_block_time = env.app.get_block_timestamp();
 
         let expected_state = StateResponse {
             is_open: true,
             is_paused: false,
-            total_staked_tokens: Uint128::new(200_000_000),
-            total_withdrawn_tokens: vec![coin(0u128, BASE_DENOM), coin(0u128, QUOTE_DENOM)],
+            total_staked_tokens: Uint128::new(250_000_000),
+            total_withdrawn_tokens: vec![coin(0u128, QUOTE_DENOM), coin(0u128, BASE_DENOM)],
             last_pause: block_time,
-            last_claim: block_time,
+            last_claim: latest_block_time,
+            pending_management_fees: vec![],
         };
 
         let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
@@ -530,4 +544,133 @@ fn test_redeem_end_to_end_with_repayment_multiple_denom() {
     );
 
     assert_approx_eq!(trader_quote_before, trader_quote_after, "2");
+}
+
+#[test]
+fn test_redeem_from_second_user() {
+    let env = TestEnv::new();
+    let wasm = Wasm::new(&env.app);
+
+    let strategy_addr = env.deploy_strategy_contract(&wasm, None);
+
+    let mut msg = env.default_fund_instantiation_msg();
+    msg.controller = strategy_addr.to_string();
+    msg.token1 = None;
+
+    let vault_addr = env.deploy_fund_contract(&wasm, msg);
+
+    env.set_vault_strategy(&wasm, &strategy_addr, &vault_addr, &env.signer)
+        .unwrap();
+    env.whitelist_hooks(vec![WhitelistedHook {
+        code_id: 2,
+        denom_creator: vault_addr.to_string(),
+    }]);
+
+    env.register_sudo_fund(&wasm, &vault_addr, &env.signer)
+        .unwrap();
+    env.set_open_fund(&wasm, &vault_addr, &env.signer).unwrap();
+
+    let config = env.query_config_fund(&wasm, &vault_addr).unwrap();
+
+    let contract_base_before = env.get_balance(&vault_addr, BASE_DENOM);
+
+    let deposit = coin(100_000_000, BASE_DENOM);
+    env.deposit_fund(&wasm, &vault_addr, &[deposit.clone()], &env.traders[0])
+        .unwrap();
+
+    let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
+    assert_eq!(state.total_staked_tokens, deposit.amount);
+
+    let trader_strategy_before = env.get_balance(&env.traders[0].address(), &config.strategy_denom);
+
+    // send funds to the trader 1
+    let redeem = coin(trader_strategy_before.u128(), config.strategy_denom.clone());
+    env.send(
+        &env.traders[1].address(),
+        redeem.clone().into(),
+        &env.traders[0],
+    );
+
+    let trader_strategy_before = env.get_balance(&env.traders[1].address(), &config.strategy_denom);
+    assert!(!trader_strategy_before.is_zero());
+
+    env.redeem_fund(&wasm, &vault_addr, redeem.clone(), &env.traders[1])
+        .unwrap();
+
+    let trader_strategy_after = env.get_balance(&env.traders[1].address(), &config.strategy_denom);
+    assert!(trader_strategy_after.is_zero());
+
+    let contract_base_after = env.get_balance(&vault_addr, BASE_DENOM);
+    assert_approx_eq!(contract_base_before, contract_base_after, "1");
+}
+
+#[test]
+fn test_redeem_from_second_user_multiple_times() {
+    let env = TestEnv::new();
+    let wasm = Wasm::new(&env.app);
+
+    let strategy_addr = env.deploy_strategy_contract(&wasm, None);
+
+    let mut msg = env.default_fund_instantiation_msg();
+    msg.controller = strategy_addr.to_string();
+    msg.token1 = None;
+
+    let vault_addr = env.deploy_fund_contract(&wasm, msg);
+
+    env.set_vault_strategy(&wasm, &strategy_addr, &vault_addr, &env.signer)
+        .unwrap();
+
+    env.whitelist_hooks(vec![WhitelistedHook {
+        code_id: 2,
+        denom_creator: vault_addr.to_string(),
+    }]);
+
+    env.register_sudo_fund(&wasm, &vault_addr, &env.signer)
+        .unwrap();
+    env.set_open_fund(&wasm, &vault_addr, &env.signer).unwrap();
+
+    let config = env.query_config_fund(&wasm, &vault_addr).unwrap();
+
+    let contract_base_before = env.get_balance(&vault_addr, BASE_DENOM);
+
+    let deposit = coin(100_000_000, BASE_DENOM);
+    env.deposit_fund(&wasm, &vault_addr, &[deposit.clone()], &env.traders[0])
+        .unwrap();
+
+    let state = env.query_state_fund(&wasm, &vault_addr).unwrap();
+    assert_eq!(state.total_staked_tokens, deposit.amount);
+
+    let trader_strategy_before = env.get_balance(&env.traders[0].address(), &config.strategy_denom);
+
+    // send funds to the trader 1
+    let send = coin(
+        trader_strategy_before.u128().checked_div(2).unwrap(),
+        config.strategy_denom.clone(),
+    );
+    env.send(
+        &env.traders[1].address(),
+        send.clone().into(),
+        &env.traders[0],
+    );
+
+    let trader_strategy_before = env.get_balance(&env.traders[0].address(), &config.strategy_denom);
+    let redeem = coin(trader_strategy_before.u128(), config.strategy_denom.clone());
+
+    env.redeem_fund(&wasm, &vault_addr, redeem.clone(), &env.traders[0])
+        .unwrap();
+
+    let trader_strategy_after = env.get_balance(&env.traders[0].address(), &config.strategy_denom);
+    assert!(trader_strategy_after.is_zero());
+
+    let trader_strategy_before = env.get_balance(&env.traders[1].address(), &config.strategy_denom);
+    let redeem = coin(trader_strategy_before.u128(), config.strategy_denom.clone());
+
+    env.redeem_fund(&wasm, &vault_addr, redeem.clone(), &env.traders[1])
+        .unwrap();
+
+    let trader_strategy_after = env.get_balance(&env.traders[1].address(), &config.strategy_denom);
+    assert!(trader_strategy_after.is_zero());
+
+    let contract_base_after = env.get_balance(&vault_addr, BASE_DENOM);
+    assert_approx_eq!(contract_base_before, contract_base_after, "1");
 }