+++ contracts/strategy/src/lib.rs
@@ -2,6 +2,7 @@ pub mod contract;
 pub mod errors;
 pub mod events;
 pub mod handle;
+pub mod ownership;
 pub mod query;
 pub mod state;
 pub mod utils;