+++ contracts/strategy/src/testing/mod.rs
@@ -1,4 +1,5 @@
 mod migration;
+mod ownership;
 mod query;
 mod repay;
 mod tests;