+++ contracts/fund/CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2024/11/12
+
+### Added
+
+- Bumped versions and integrated with `cosmwasm-std@v2.0.0`
+
+## [0.0.6] - 2024/11/06
+
+### Added
+
+- Concept of management fee, that is paid to the treasury contract on each contract interaction
+
 ## [0.0.5] - 2024/10/17
 
 ### Fixed